chore: move PreferencesSection to its own folder
