+++ src/client/v2/src/components/SettingsModal/sections/PreferencesSection/PreferencesSection.test.tsx
